+++ e2e/imports.ts
@@ -1,4 +1,4 @@
-import { readJSONSync } from 'fs-extra';
+import { readFileSync } from 'node:fs';
 
 /**
  * The import map used by `./utils/setupApp` to inject into the page so test scripts can use modules.
@@ -7,8 +7,9 @@ import { readJSONSync } from 'fs-extra';
  *
  * **IMPORTANT**: be sure to update the paths field in `./tsconfig.json` to reflect imports correctly
  */
+const packageJSON = JSON.parse(readFileSync('./package.json', 'utf8'));
+
 export default {
-  fabric: readJSONSync('./package.json').module.slice(1),
-  ['fabric/extensions']:
-    readJSONSync('./package.json').exports['./extensions'].import.slice(1),
+  fabric: packageJSON.module.slice(1),
+  ['fabric/extensions']: packageJSON.exports['./extensions'].import.slice(1),
 };